added a hintset object to the spoiler, removing global var
