+++ tests/test_spoiler.py
@@ -36,7 +36,7 @@ class TestSpoiler(unittest.TestCase):
     def test_settings_string(self):
         """Confirm that settings strings decryption is working and generate a spoiler log with it."""
         # The settings string is defined either here or in the string above, pick your poison
-        self.settings_string = "Px+YQAIq9AANfgACwQABYYAEMUAAH+gaZAUBAwEBwMEAgJBQUDAsHBgRAA6gA3AA3gA4AA4gA5AA5gA6ABcgCAhOogQoVAwwlggTVACtAtEtItMtRNVNZtdti2a2obbbjbqd1bzcJcBb4d9eFcQc5cYcicrcwdAdJdUdadjdsNGDRJJqTMLjAMtUSorIEFECjBSApQUwHuZ7JsoIASNFaCCLgnqbJ8s2zII1hZZD9d2PDISytgLcHIymKaQM5Oz8izUkvRYhbF6p0KiFgEJiFAEkSO1VVZQMZYcmSd4igOIZXkCsZ5LGRkAMqmeIboKQgAFoUAC0MABKHAAlEAAOrQAHRIACooAA0WAA6MAAVGgAGrgALSMnolkwVRdFszxZBgjhgJ8DhTFQtBWBoKxDEIawuJ4ST+QcliDLovB0IChAREhQVFhgZGhwdHiAhIiMkJSYnKCksLS4vMF8DBIWGBodWMADXF1eYQQVFZiZGZoar7EADsAP0BKcCgxGQx0JW5Oh4kaKzxNXm94dI8YBDuuIMtYhBevn8A2ePhhNHO7qV7KzM4ps6ti+FOFAMCRKlsa58IzP4AsgCCgBRHQiGQpjoSpbGufCMp0ra+M63zryAj1HuCSoIqvMSnmCDFKJMSWYacehopifAkCIMOQcamuy2+7BHJUW9AB8gD1AH0APYAfYA"
+        self.settings_string = "PyEAAAvQADX4ADMEAAmGAADFAAB/oGmQFAQMBAcDBAICQUFAwLBwYEQBOqWNwAN4AOAAOIAOQAOYAOgAXIAJmAhOegQoTAwwmhIJrIBWgWiWkWmWomqms2u2xbNbUNttxt1O6t5uEuAt8O+vCuIOcuMOROVuYOgOkuqOtOxu2GiJJNSZRcYB1qiVAIKIFGCkBSgpgIIuCepsnyzbMQjZZD9PwyEsrYC3ASMximkCOj4hTUkvRYhbF6p0KiFgEJmFEAkSO1VVZQMZYcmSd4igOV5ArHB55LGBkAAqmeIboIQgAFoUAC0MABKHAAlEAAOrQAHRIACooAA0WAA6MAAVGgAGrgALJ8Jb8FVBisF0Wx7M8WQYI4YBnMksC7A4UxULQxxWBoKxDEUawuJ4SQhP4liDJEui8IggJCgsNDhAREhQVFhgZGhwdHiAhIiMkJSYnKCksLS4vMDVfAwYVmJov1jAA1xdXmEBlRkZsQAOwA/QEpwJDEZDpWx6dDxI0VniavN7w6R4wCHdcQZaxCC9fP4Bs8fDCaOd3Ur2VmZxTZ1bF8KcKAeEQyGQlS2Nc+ET+ALIAgoG1CiPBEMhTHQlS2Nc+EZTpW18Z1vnfPIGsmnUioe4A8w8IMUkow9JZhpx6Gl8C4Igw5BZpyCKaq7Lbr8EcktVb0AHyAPUAfQA9gB9gA"
         settings_dict = decrypt_settings_string_enum(self.settings_string)
         settings_dict["seed"] = random.randint(0, 100000000)  # Can be fixed if you want to test a specific seed repeatedly
